評価機能の実装

+++ recipe/src/main/java/com/example/recipe/dto/RecipeListDto.java
@@ -28,7 +28,7 @@ public class RecipeListDto implements Serializable {
 	/**
 	 * 評価
 	 */
-	private int recipeAveRating;
+	private double recipeAveRating;
 	
 	/*
 	 * レシピ画像
